+++ modules/back-end/src/Application/ConfigurationExtensions.cs
@@ -1,0 +1,13 @@
+using Microsoft.Extensions.Configuration;
+
+namespace Application;
+
+public static class ConfigurationExtensions
+{
+    public static bool IsSaasHosting(this IConfiguration configuration)
+    {
+        var mode = configuration.GetSection(HostingMode.SectionName).Value ?? HostingMode.SelfHosted;
+
+        return mode == HostingMode.SaaS;
+    }
+}
