+++ modules/back-end/src/Application/HostingMode.cs
@@ -1,4 +1,4 @@
-namespace Infrastructure;
+namespace Application;
 
 public static class HostingMode
 {
