+++ modules/back-end/src/Application/Workspaces/UpdateLicense.cs
@@ -29,39 +29,29 @@ public class UpdateLicenseValidator : AbstractValidator<UpdateLicense>
     }
 }
 
-public class UpdateLicenseHandler : IRequestHandler<UpdateLicense, WorkspaceVm>
+public class UpdateLicenseHandler(
+    IWorkspaceService service,
+    ICacheService cacheService,
+    IMapper mapper,
+    IConfiguration configuration)
+    : IRequestHandler<UpdateLicense, WorkspaceVm>
 {
-    private readonly IWorkspaceService _service;
-    private readonly ICacheService _cacheService;
-    private readonly IMapper _mapper;
-    private readonly bool _isSaas;
-
-    public UpdateLicenseHandler(IWorkspaceService service, ICacheService cacheService, IMapper mapper, IConfiguration configuration)
-    {
-        _service = service;
-        _cacheService = cacheService;
-        _mapper = mapper;
-
-        var hostingMode = configuration.GetSection("HostingMode").Value;
-        _isSaas = hostingMode == "saas";
-    }
-
     public async Task<WorkspaceVm> Handle(UpdateLicense request, CancellationToken cancellationToken)
     {
-        if (_isSaas)
+        if (configuration.IsSaasHosting())
         {
-            throw new BusinessException("Forbidden");
+            throw new BusinessException(ErrorCodes.Forbidden);
         }
 
-        var workspace = await _service.GetAsync(request.Id);
+        var workspace = await service.GetAsync(request.Id);
 
         // save to database
         workspace.UpdateLicense(request.License);
-        await _service.UpdateAsync(workspace);
+        await service.UpdateAsync(workspace);
 
         // update license cache
-        await _cacheService.UpsertLicenseAsync(workspace);
+        await cacheService.UpsertLicenseAsync(workspace);
 
-        return _mapper.Map<WorkspaceVm>(workspace);
+        return mapper.Map<WorkspaceVm>(workspace);
     }
 }
