+++ modules/back-end/src/Infrastructure/ConfigureServices.cs
@@ -1,6 +1,6 @@
+using Application;
 using Application.Usages;
 using Domain.Users;
-using Infrastructure;
 using Infrastructure.Caches;
 using Infrastructure.MQ;
 using Infrastructure.Persistence;
@@ -52,7 +52,7 @@ public static class ConfigureServices
         services.AddDbSpecificServices(configuration);
         services.AddTransient<IEnvironmentAppService, AppServices.EnvironmentAppService>();
         services.AddTransient<IFeatureFlagAppService, AppServices.FeatureFlagAppService>();
-        if (configuration.GetHostingMode() == HostingMode.SaaS)
+        if (configuration.IsSaasHosting())
         {
             services.AddTransient<IBillingService, Services.BillingService>();
         }
